13608 & 12026 - align regular and CDC integration tests and data mappers
