+++ airbyte-integrations/bases/debezium-v1-9-2/src/main/java/io/airbyte/integrations/debezium/internals/PostgresConverter.java
@@ -4,12 +4,26 @@
 
 package io.airbyte.integrations.debezium.internals;
 
+import static io.airbyte.protocol.models.JsonSchemaType.AIRBYTE_TYPE;
+import static io.airbyte.protocol.models.JsonSchemaType.DATE;
+import static io.airbyte.protocol.models.JsonSchemaType.DATE_TIME;
+import static io.airbyte.protocol.models.JsonSchemaType.FORMAT;
+import static io.airbyte.protocol.models.JsonSchemaType.TIME;
+import static io.airbyte.protocol.models.JsonSchemaType.TIMESTAMP_WITHOUT_TIMEZONE;
+import static io.airbyte.protocol.models.JsonSchemaType.TIMESTAMP_WITH_TIMEZONE;
+import static io.airbyte.protocol.models.JsonSchemaType.TIME_WITHOUT_TIMEZONE;
+import static io.airbyte.protocol.models.JsonSchemaType.TIME_WITH_TIMEZONE;
+
 import io.debezium.spi.converter.CustomConverter;
 import io.debezium.spi.converter.RelationalColumn;
 import java.math.BigDecimal;
 import java.nio.charset.StandardCharsets;
+import java.sql.Date;
+import java.sql.Timestamp;
 import java.util.Arrays;
+import java.util.Locale;
 import java.util.Properties;
+import org.apache.commons.codec.binary.Hex;
 import org.apache.kafka.connect.data.SchemaBuilder;
 import org.postgresql.util.PGInterval;
 import org.slf4j.Logger;
@@ -19,12 +33,14 @@ public class PostgresConverter implements CustomConverter<SchemaBuilder, Relatio
 
   private static final Logger LOGGER = LoggerFactory.getLogger(PostgresConverter.class);
 
-  private final String[] DATE_TYPES = {"DATE", "DATETIME", "TIME", "TIMETZ", "INTERVAL", "TIMESTAMP"};
+  private final String[] DATE_TYPES = {"DATE", "DATETIME", "TIME", "TIMETZ", "INTERVAL", "TIMESTAMP", "TIMESTAMPTZ"};
   private final String[] BIT_TYPES = {"BIT", "VARBIT"};
   private final String[] MONEY_ITEM_TYPE = {"MONEY"};
   private final String[] GEOMETRICS_TYPES = {"BOX", "CIRCLE", "LINE", "LSEG", "POINT", "POLYGON", "PATH"};
   private final String[] TEXT_TYPES =
-      {"VARCHAR", "VARBINARY", "BLOB", "TEXT", "LONGTEXT", "TINYTEXT", "MEDIUMTEXT", "INVENTORY_ITEM", "TSVECTOR", "TSQUERY"};
+      {"VARCHAR", "VARBINARY", "BLOB", "TEXT", "LONGTEXT", "TINYTEXT", "MEDIUMTEXT", "INVENTORY_ITEM", "TSVECTOR", "TSQUERY", "PG_LSN"};
+  private final String[] NUMERIC_TYPES = {"NUMERIC", "DECIMAL"};
+  private final String BYTEA_TYPE = "BYTEA";
 
   @Override
   public void configure(final Properties props) {}
@@ -39,7 +55,48 @@ public class PostgresConverter implements CustomConverter<SchemaBuilder, Relatio
       registerText(field, registration);
     } else if (Arrays.stream(MONEY_ITEM_TYPE).anyMatch(s -> s.equalsIgnoreCase(field.typeName()))) {
       registerMoney(field, registration);
+    } else if (BYTEA_TYPE.equalsIgnoreCase(field.typeName())) {
+      registerBytea(field, registration);
     }
+    else if (Arrays.stream(NUMERIC_TYPES).anyMatch(s -> s.equalsIgnoreCase(field.typeName()))) {
+      registerNumber(field, registration);
+    }
+  }
+
+  private void registerNumber(final RelationalColumn field, final ConverterRegistration<SchemaBuilder> registration) {
+    registration.register(SchemaBuilder.string().optional(), x -> {
+      if (x == null) {
+        return DebeziumConverterUtils.convertDefaultValue(field);
+      }
+//      Bad solution
+//      We applied a solution like this for several reasons:
+//      1. Regarding #13608, CDC and nor-CDC data output format should be the same.
+//      2. In the non-CDC mode 'decimal' and 'numeric' values are put to JSON node as BigDecimal value.
+//      According to Jackson Object mapper configuration, all trailing zeros are omitted and
+//      numbers with decimal places are deserialized with exponent. (e.g. 1234567890.1234567 would
+//      be deserialized as 1.2345678901234567E9).
+//      3. In the CDC mode 'decimal' and 'numeric' values are deserialized as a regular number (e.g.
+//      1234567890.1234567 would be deserialized as 1234567890.1234567). Numbers without
+//      decimal places (e.g 1, 24, 354) are represented with trailing zero (e.g 1.0, 24.0, 354.0).
+//      One of solution to align deserialization for these 2 modes is setting
+//      DeserializationFeature.USE_BIG_DECIMAL_FOR_FLOATS as true for ObjectMapper. But this breaks
+//      deserialization for other data-types.
+//      A worked solution was to keep deserialization for non-CDC mode as it is and change it for CDC one.
+//      The code below strips trailing zeros for integer numbers and represents number with exponent
+//      if this number has decimals point.
+      final double doubleValue = Double.parseDouble(x.toString());
+      var valueWithTruncatedZero = BigDecimal.valueOf(doubleValue).stripTrailingZeros().toString();
+      return valueWithTruncatedZero.contains(".") ? String.valueOf(doubleValue) : valueWithTruncatedZero;
+    });
+  }
+
+  private void registerBytea(final RelationalColumn field, final ConverterRegistration<SchemaBuilder> registration) {
+    registration.register(SchemaBuilder.string().optional(), x -> {
+      if (x == null) {
+        return DebeziumConverterUtils.convertDefaultValue(field);
+      }
+      return "\\x" + Hex.encodeHexString((byte[]) x);
+    });
   }
 
   private void registerText(final RelationalColumn field, final ConverterRegistration<SchemaBuilder> registration) {
@@ -57,14 +114,23 @@ public class PostgresConverter implements CustomConverter<SchemaBuilder, Relatio
   }
 
   private void registerDate(final RelationalColumn field, final ConverterRegistration<SchemaBuilder> registration) {
+    final var fieldType = field.typeName();
+
     registration.register(SchemaBuilder.string().optional(), x -> {
       if (x == null) {
         return DebeziumConverterUtils.convertDefaultValue(field);
-      } else if (x instanceof PGInterval) {
-        return convertInterval((PGInterval) x);
-      } else {
-        return DebeziumConverterUtils.convertDate(x);
       }
+      return switch (fieldType.toUpperCase(Locale.ROOT)) {
+        case "TIMETZ" -> DateTimeConverter.convertToTimeWithTimezone(x);
+        case "TIMESTAMPTZ" -> DateTimeConverter.convertToTimestampWithTimezone(x);
+        case "TIMESTAMP" -> DateTimeConverter.convertToTimestamp(x);
+        // Debezium doesn't handle era indicators
+        // https://github.com/airbytehq/airbyte/issues/14590
+        case "DATE" -> DateTimeConverter.convertToDate(x);
+        case "TIME" -> DateTimeConverter.convertToTime(x);
+        case "INTERVAL" -> convertInterval((PGInterval) x);
+        default -> DebeziumConverterUtils.convertDate(x);
+      };
     });
   }
 
@@ -84,11 +150,7 @@ public class PostgresConverter implements CustomConverter<SchemaBuilder, Relatio
         return DebeziumConverterUtils.convertDefaultValue(field);
       } else if (x instanceof Double) {
         final BigDecimal result = BigDecimal.valueOf((Double) x);
-        if (result.compareTo(new BigDecimal("999999999999999")) == 1
-            || result.compareTo(new BigDecimal("-999999999999999")) == -1) {
-          return null;
-        }
-        return result.toString();
+        return Double.toString(result.doubleValue());
       } else {
         return x.toString();
       }
