+++ airbyte-integrations/connectors/source-postgres/src/main/java/io/airbyte/integrations/source/postgres/PostgresSourceOperations.java
@@ -23,6 +23,7 @@ import io.airbyte.db.DataTypeUtils;
 import io.airbyte.db.jdbc.JdbcSourceOperations;
 import io.airbyte.protocol.models.JsonSchemaType;
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.sql.JDBCType;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -35,7 +36,15 @@ import java.time.OffsetDateTime;
 import java.time.OffsetTime;
 import java.time.format.DateTimeParseException;
 import java.util.Collections;
+import org.postgresql.geometric.PGbox;
+import org.postgresql.geometric.PGcircle;
+import org.postgresql.geometric.PGline;
+import org.postgresql.geometric.PGlseg;
+import org.postgresql.geometric.PGpath;
+import org.postgresql.geometric.PGpoint;
+import org.postgresql.geometric.PGpolygon;
 import org.postgresql.jdbc.PgResultSetMetaData;
+import org.postgresql.util.PGobject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -173,6 +182,14 @@ public class PostgresSourceOperations extends JdbcSourceOperations {
         case TIMETZ -> putTimeWithTimezone(json, columnName, resultSet, colIndex);
         case TIMESTAMPTZ -> putTimestampWithTimezone(json, columnName, resultSet, colIndex);
         case "hstore" -> putHstoreAsJson(json, columnName, resultSet, colIndex);
+        case "circle" -> putObject(json, columnName, resultSet, colIndex, PGcircle.class);
+        case "box" -> putObject(json, columnName, resultSet, colIndex, PGbox.class);
+        case "double precision", "float", "float8" -> putFloat8(json, columnName, resultSet, colIndex);
+        case "line" -> putObject(json, columnName, resultSet, colIndex, PGline.class);
+        case "lseg" -> putObject(json, columnName, resultSet, colIndex, PGlseg.class);
+        case "path" -> putObject(json, columnName, resultSet, colIndex, PGpath.class);
+        case "point" -> putObject(json, columnName, resultSet, colIndex, PGpoint.class);
+        case "polygon" -> putObject(json, columnName, resultSet, colIndex, PGpolygon.class);
         default -> {
           switch (columnType) {
             case BOOLEAN -> putBoolean(json, columnName, resultSet, colIndex);
@@ -198,19 +215,19 @@ public class PostgresSourceOperations extends JdbcSourceOperations {
 
   @Override
   protected void putDate(final ObjectNode node, final String columnName, final ResultSet resultSet, final int index) throws SQLException {
-    final LocalDate date = getDateTimeObject(resultSet, index, LocalDate.class);
+    final LocalDate date = getObject(resultSet, index, LocalDate.class);
     node.put(columnName, resolveEra(date, date.toString()));
   }
 
   @Override
   protected void putTime(final ObjectNode node, final String columnName, final ResultSet resultSet, final int index) throws SQLException {
-    final LocalTime time = getDateTimeObject(resultSet, index, LocalTime.class);
+    final LocalTime time = getObject(resultSet, index, LocalTime.class);
     node.put(columnName, time.format(TIME_FORMATTER));
   }
 
   @Override
   protected void putTimestamp(final ObjectNode node, final String columnName, final ResultSet resultSet, final int index) throws SQLException {
-    final LocalDateTime timestamp = getDateTimeObject(resultSet, index, LocalDateTime.class);
+    final LocalDateTime timestamp = getObject(resultSet, index, LocalDateTime.class);
     final LocalDate date = timestamp.toLocalDate();
     node.put(columnName, resolveEra(date, timestamp.format(TIMESTAMP_FORMATTER)));
   }
@@ -264,6 +281,26 @@ public class PostgresSourceOperations extends JdbcSourceOperations {
   }
 
   @Override
+  protected void putFloat8(final ObjectNode node, final String columnName, final ResultSet resultSet, final int index)
+      throws SQLException {
+    final BigDecimal bigDecimal = DataTypeUtils.returnNullIfInvalid(() -> resultSet.getBigDecimal(index));
+    if (bigDecimal != null) {
+      node.put(columnName, bigDecimal.setScale(resultSet.getMetaData().getScale(index), RoundingMode.HALF_EVEN).doubleValue());
+    } else {
+      node.put(columnName, (BigDecimal) null);
+    }
+  }
+
+  protected <T extends PGobject> void putObject(final ObjectNode node,
+                                                final String columnName,
+                                                final ResultSet resultSet,
+                                                final int index,
+                                                Class<T> clazz)
+      throws SQLException {
+    final T object = getObject(resultSet, index, clazz);
+    node.put(columnName, object.getValue());
+  }
+
   protected void putBigDecimal(final ObjectNode node, final String columnName, final ResultSet resultSet, final int index) {
     final BigDecimal bigDecimal = DataTypeUtils.returnNullIfInvalid(() -> resultSet.getBigDecimal(index));
     if (bigDecimal != null) {
